Refactor semantic module to decouple data, logic, and types.
- Created `src/semantic/model.rs` for AST definitions (`AnalyzedStatement`, `TraitDef`, etc.)
- Cleaned `src/semantic/types.rs` to only contain Type System definitions (`GlossaType`).
- Cleaned `src/semantic/mod.rs` by removing legacy code (`SemanticAnalyzer`) and re-exporting the model.
- Moved `AnalyzedProgram` to `src/semantic/mod.rs` to prevent circular dependencies.
- Updated imports in `resolver.rs` and `declarations.rs`.
- Removed dead code `classify_value_expression`.

This change improves architectural modularity and eliminates the need for legacy analysis paths.
