+++ .jules/atlas.md
@@ -10,3 +10,11 @@
 ## [Splitting The Blob: Semantic Module]
 **Tangle:** `src/semantic/mod.rs` had grown to over 4600 lines, mixing high-level orchestration, control flow parsing, pattern matching, and expression analysis. This violated the Single Responsibility Principle and made navigation difficult.
 **Blueprint:** Split `src/semantic/mod.rs` into `control_flow.rs`, `declarations.rs`, `expressions.rs`, `patterns.rs`, and `conversion.rs`. `mod.rs` now acts as an orchestrator and facade.
+
+## [Decoupling Semantic Model]
+**Tangle:** The `semantic` module suffered from poor separation of concerns. Data structures like `AnalyzedProgram` and `TraitDef` were scattered between `mod.rs` and `types.rs`. `mod.rs` contained legacy analysis logic (`SemanticAnalyzer`) mixed with module orchestration. `types.rs` contained both Type System definitions (`GlossaType`) and AST nodes (`TraitDef`), creating implicit circular dependencies.
+**Blueprint:**
+1. Created `src/semantic/model.rs` to house all Semantic AST nodes (`AnalyzedStatement`, `TraitDef`, etc.).
+2. Purified `src/semantic/types.rs` to contain only Type System definitions (`GlossaType`).
+3. Cleaned `src/semantic/mod.rs` by removing legacy code and re-exporting the model.
+**Stability:** Separates "Data" (model) from "Logic" (analysis) and "Types" (type system). Eliminates potential circular dependencies and makes the module structure intuitive.
