+++ src/semantic/conversion.rs
@@ -26,98 +26,6 @@ pub fn convert_assembled_to_analyzed(
     Ok(AnalyzedStatement { kind, expressions })
 }
 
-/// Convert an AssembledStatement to a single AnalyzedExpr for use in value expressions
-/// Handles patterns like: ξ ἓν ἄθροισμα → BinOp(Variable("xi"), Add, NumberLiteral(1))
-pub fn classify_value_expression(
-    asm_stmt: &AssembledStatement,
-    scope: &mut Scope,
-) -> Result<AnalyzedExpr, GlossaError> {
-    // Check for binary operation: subject + object + operator
-    if let Some(subject) = &asm_stmt.subject {
-        if !asm_stmt.operators.is_empty() {
-            let subj_name = normalize_greek(&subject.original);
-            let left = AnalyzedExpr {
-                expr: AnalyzedExprKind::Variable(subj_name.clone()),
-                glossa_type: scope
-                    .lookup(&subj_name)
-                    .cloned()
-                    .unwrap_or(GlossaType::Number),
-            };
-
-            // Get the right operand (from literals or object)
-            let right = if !asm_stmt.literals.is_empty() {
-                match &asm_stmt.literals[0] {
-                    crate::semantic::assembler::Literal::Number(n) => AnalyzedExpr {
-                        expr: AnalyzedExprKind::NumberLiteral(*n),
-                        glossa_type: GlossaType::Number,
-                    },
-                    crate::semantic::assembler::Literal::String(s) => AnalyzedExpr {
-                        expr: AnalyzedExprKind::StringLiteral(s.clone()),
-                        glossa_type: GlossaType::String,
-                    },
-                    crate::semantic::assembler::Literal::Boolean(b) => AnalyzedExpr {
-                        expr: AnalyzedExprKind::BooleanLiteral(*b),
-                        glossa_type: GlossaType::Boolean,
-                    },
-                }
-            } else if let Some(object) = &asm_stmt.object {
-                let obj_name = normalize_greek(&object.original);
-                AnalyzedExpr {
-                    expr: AnalyzedExprKind::Variable(obj_name.clone()),
-                    glossa_type: scope
-                        .lookup(&obj_name)
-                        .cloned()
-                        .unwrap_or(GlossaType::Number),
-                }
-            } else {
-                return Err(GlossaError::semantic(
-                    "Binary operation missing right operand",
-                ));
-            };
-
-            let op = asm_stmt.operators[0];
-            return Ok(AnalyzedExpr {
-                expr: AnalyzedExprKind::BinOp {
-                    left: Box::new(left),
-                    op,
-                    right: Box::new(right),
-                },
-                glossa_type: GlossaType::Number,
-            });
-        }
-
-        // Just a variable reference
-        let var_name = normalize_greek(&subject.original);
-        return Ok(AnalyzedExpr {
-            expr: AnalyzedExprKind::Variable(var_name.clone()),
-            glossa_type: scope
-                .lookup(&var_name)
-                .cloned()
-                .unwrap_or(GlossaType::Number),
-        });
-    }
-
-    // Check for literal-only value
-    if !asm_stmt.literals.is_empty() {
-        return match &asm_stmt.literals[0] {
-            crate::semantic::assembler::Literal::Number(n) => Ok(AnalyzedExpr {
-                expr: AnalyzedExprKind::NumberLiteral(*n),
-                glossa_type: GlossaType::Number,
-            }),
-            crate::semantic::assembler::Literal::String(s) => Ok(AnalyzedExpr {
-                expr: AnalyzedExprKind::StringLiteral(s.clone()),
-                glossa_type: GlossaType::String,
-            }),
-            crate::semantic::assembler::Literal::Boolean(b) => Ok(AnalyzedExpr {
-                expr: AnalyzedExprKind::BooleanLiteral(*b),
-                glossa_type: GlossaType::Boolean,
-            }),
-        };
-    }
-
-    Err(GlossaError::semantic("Unable to classify value expression"))
-}
-
 /// Classify an assembled statement and extract analyzed expressions
 pub fn classify_assembled_statement(
     asm_stmt: &AssembledStatement,
